files renamed, admin entity deleted, controllers and services adapted.
