+++ mcsv-administrator/pom.xml
@@ -8,22 +8,15 @@
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.arqui.integrador</groupId>
-	<artifactId>mcsv-administrator</artifactId>
+	<artifactId>mcsv-admin</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>mcsv-administrator</name>
+	<name>mcsv-admin</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
 		<java.version>17</java.version>
 		<spring-cloud.version>2022.0.4</spring-cloud.version>
 	</properties>
 	<dependencies>
-		
-		<dependency>
-			<groupId>org.springdoc</groupId>
-			<artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
-			<version>2.2.0</version>
-		 </dependency>
-	  
 
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -54,8 +47,8 @@
 		</dependency>
 
 		<dependency>
-			<groupId>com.mysql</groupId>
-			<artifactId>mysql-connector-j</artifactId>
+			<groupId>org.postgresql</groupId>
+			<artifactId>postgresql</artifactId>
 			<scope>runtime</scope>
 		</dependency>
