+++ mcsv-administrator/src/main/java/com/arqui/integrador/mcsvadmin/McsvAdminApplication.java
@@ -1,4 +1,4 @@
-package com.arqui.integrador.mcsvadministrator;
+package com.arqui.integrador.mcsvadmin;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,9 +6,9 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-public class McsvAdministratorApplication {
+public class McsvAdminApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(McsvAdministratorApplication.class, args);
+		SpringApplication.run(McsvAdminApplication.class, args);
 	}
 }
