+++ mcsv-administrator/src/main/java/com/arqui/integrador/mcsvadmin/config/AdminConfig.java
@@ -1,4 +1,4 @@
-package com.arqui.integrador.mcsvadministrator.config;
+package com.arqui.integrador.mcsvadmin.config;
 
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
@@ -6,11 +6,11 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class AdministratorConfig {
+public class AdminConfig {
     
     @Bean
     @LoadBalanced
-	public RestTemplate restTemplateAdministrator() { 
+	public RestTemplate restTemplateAdmin() { 
         return new RestTemplate();
     }
 }
