+++ mcsv-administrator/src/main/java/com/arqui/integrador/mcsvadmin/dto/ScooterForMaintenanceDto.java
@@ -1,0 +1,5 @@
+package com.arqui.integrador.mcsvadmin.dto;
+
+public class ScooterForMaintenanceDto {
+    
+}
