+++ mcsv-administrator/src/main/java/com/arqui/integrador/mcsvadmin/dto/TravelsByTotalBillingAmountDto.java
@@ -1,4 +1,4 @@
-package com.arqui.integrador.mcsvadministrator.dto;
+package com.arqui.integrador.mcsvadmin.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 
-public class TravelsByTotalBillingAmount {
+public class TravelsByTotalBillingAmountDto {
 
     private Double total;
