+++ mcsv-administrator/src/main/java/com/arqui/integrador/mcsvadmin/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.arqui.integrador.mcsvadministrator.exception;
+package com.arqui.integrador.mcsvadmin.exception;
 
 import java.sql.Timestamp;
