--- mcsv-administrator/src/main/java/com/arqui/integrador/mcsvadministrator/dto/ScooterForMaintenanceDTO.java
@@ -1,5 +1,0 @@
-package com.arqui.integrador.mcsvadministrator.dto;
-
-public class ScooterForMaintenanceDTO {
-    
-}
