--- mcsv-administrator/src/main/java/com/arqui/integrador/mcsvadministrator/model/Administrator.java
@@ -1,36 +1,0 @@
-package com.arqui.integrador.mcsvadministrator.model;
-
-import java.io.Serializable;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "app_admin")
-public class Administrator implements Serializable {
-   
-	private static final long serialVersionUID = 3253615484910702243L;
-
-	@Id
-    @Column(name = "admin_id")
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    private String name;
-
-    private String password;
-
-    private String rol; // TODO pasarlo a enum
-}
