+++ mcsv-administrator/src/test/java/com/arqui/integrador/mcsvadmin/McsvAdministratorApplicationTests.java
@@ -1,4 +1,4 @@
-package com.arqui.integrador.mcsvadministrator;
+package com.arqui.integrador.mcsvadmin;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
